backend/requirements.txt
APScheduler==3.10.4
anthropic
asyncpg==0.30.0
beautifulsoup4==4.12.3
elasticsearch
fastapi==0.115.12
google-generativeai==0.8.5
httpx==0.28.1
opensearch-py==2.6.0
pydantic
python-jose
qiskit==1.1.2
qiskit-aer==0.14.2
qiskit-ibm-runtime==0.22.0
redis==5.2.1
uvicorn==0.34.0
